Move examples to demo app
This moves the three existing examples to the consolidated demo app
