+++ demo/src/GitHubRestApi/create-store.js
@@ -1,8 +1,10 @@
 import { createStore } from 'redux';
 
-export default ( reducer ) => {
+const createGitHubRestApiStore = ( reducer ) => {
 	return createStore(
 		reducer,
 		window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 	);
 };
+
+export default createGitHubRestApiStore;
