+++ demo/src/GitHubRestApi/reducer.js
