+++ demo/src/WPRestApi/create-store.js
@@ -1,8 +1,10 @@
 import { createStore } from 'redux';
 
-export default ( reducer ) => {
+const createWPApiStore = ( reducer ) => {
 	return createStore(
 		reducer,
 		window.__REDUX_DEVTOOLS_EXTENSION__ && window.__REDUX_DEVTOOLS_EXTENSION__()
 	);
 };
+
+export default createWPApiStore;
