+++ demo/src/WPRestApi/reducer.js
