--- examples/hello-world/.env
@@ -1,3 +1,0 @@
-# This is necessary because react-scripts wants an old version of babel-eslint.
-# After react-scripts updates to v10.x.x this can be removed.
-SKIP_PREFLIGHT_CHECK=true
