--- examples/hello-world/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "example_hello-world",
-  "version": "0.7.0",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  },
-  "devDependencies": {
-    "react-scripts": "^3.4.4"
-  },
-  "dependencies": {
-    "@fresh-data/framework": "^0.7.0",
-    "@fresh-data/react-provider": "^0.7.0",
-    "ajv": "^6.12.6",
-    "debug": "^4.3.1",
-    "lodash": "^4.17.20",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^6.0.1",
-    "redux": "^4.0.5"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
-}
