--- examples/hello-world/public/manifest.json
@@ -1,8 +1,0 @@
-{
-  "short_name": "Fresh Data",
-  "name": "Fresh Data Example - Hello World",
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
