--- examples/hello-world/src/App.css
@@ -1,17 +1,0 @@
-.App {
-  text-align: center;
-}
-
-.App-header {
-  background-color: rgb(57, 184, 99);
-  padding: 20px;
-  color: white;
-}
-
-.App-title {
-  font-size: 1.5em;
-}
-
-.App-intro {
-  font-size: large;
-}
