--- examples/wp-rest-api/README.md
@@ -1,11 +1,0 @@
-# Fresh Data Example - WordPress REST API
-
-This example code shows how to use data from public WordPress API calls.
-
-This example project was built with [Create React App](https://github.com/facebook/create-react-app).
-
-## Try it out!
---------------
-Simply run:
-
-```npm start```
