--- examples/wp-rest-api/public/manifest.json
@@ -1,8 +1,0 @@
-{
-  "short_name": "Fresh Data - WP REST API",
-  "name": "Fresh Data Example - WordPress REST API",
-  "start_url": "./index.html",
-  "display": "standalone",
-  "theme_color": "#000000",
-  "background_color": "#ffffff"
-}
